Add option to set driver in ImpalaJdbcDriver
ImpalaJdbcDriver is a simple class that provides a wrapper around the
Hive Jdbc driver for use in testing. This patch adds a '-d' parameter
that takes a driver class and executes queries with that driver
instead of the Hive driver, if its in the classpath.

This makes it easy to use ./bin/run-jdbc-client.sh to test other
Jdbc drivers, such as the proprietary Cloudera Impala driver.

Change-Id: Ie6c8cec2f61e76b7c8321c954eef830778d8a8ee