CDPD-10316: Add RAZ JARS to Impala's classpath
Hadoop starts to support Shared Access Signatures (SAS) after
HADOOP-16730, which allows an Azure Blob File System (ABFS) client to
use temporary tokens with constrained access to a specific file or
directory. The interface SASTokenProvider defines a method getSASToken()
that retrieves such SAS tokens and the class RangerRazTokenProvider
implements this interface.

On the other hand, recall that the method
FileSystemUtil#getTotalNumVisibleFiles(Path directory) in Impala calls
FileSystem#getFileStatus() to check the given Path is indeed a
directory. When the FileSystem is an instance of AzureBlobFileSystem,
AbfsClient#getAclStatus() will be called, which in turn invokes
getSASToken(). This patch adds the required dependency so as to use
getSASToken() implemented by RangerRazTokenProvider.

Testing:
- Verified that the jar files of ranger-raz-hook-abfs and
  ranger-raz-intg are downloaded to the m2 directory after building
  Impala in a development environment.

Change-Id: Iafc9d17c8216ce3a6f9884a34f3defae2312838e
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/78131
CDH-Build: Jenkins User <jenkins@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>
Quasar-L0: quasar-precommit User
Reviewed-by: Madhan Neethiraj <mneethiraj@cloudera.com>
Reviewed-by: Fang-Yu Rao <fangyu.rao@cloudera.com>