CDPD-10316: Add RAZ jar files to Impala's classpath
Hadoop starts to support Shared Access Signatures (SAS) after
HADOOP-16730, which allows an Azure Blob File System (ABFS) client to
use a temporary token with constrained access to a specific file or
directory. The interface SASTokenProvider defines a method getSASToken()
that retrieves such SAS tokens and the class RangerRazTokenProvider
implements this interface.

On the other hand, recall that the method
FileSystemUtil#getTotalNumVisibleFiles(Path directory) in Impala calls
FileSystem#getFileStatus() to check the given Path is indeed a
directory. When the FileSystem is an instance of AzureBlobFileSystem,
AbfsClient#getAclStatus() will be called, which in turn invokes
getSASToken(). This patch adds the required dependency so as to use
getSASToken() implemented by RangerRazTokenProvider.

Testing:
- Verified that the jar files of ranger-raz-hook-abfs and
  ranger-raz-intg are downloaded to the m2 directory after building
  Impala in a development environment.
- Examined fe/pom.xml to make sure the newly included jar files do not
  contain any jar file that should be excluded, e.g., jersey-bundle.
- Verified that this patch passes the core tests in the DEBUG build
  except for 6 unrelated E2E tests.

Change-Id: I049002bdd3faf258526786bcbbbc6e82664a4611
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/78805
CDH-Build: Jenkins User <jenkins@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Quasar-L0: quasar-precommit User