CDPD-10444 (Impala side): Update the version of Atlas' Impala hook
Due to the incompatibility of jackson-databind reported at CDPD-10218,
at the end of a query execution in Impala, we saw an error of
NoClassDefFoundError resulting from the call to
QueryEventHookManager#executeQueryCompleteHooks(). Now that the issue
has been resolved on the Atlas side, Impala should use an upgraded hook
provided by Atlas.

To address this issue, instead of using a hook of a hard-coded version
that has the fix as previously done, this patch enables Impala's build
script to always download the up-to-date hook in either the standalone
build or the CDP build.

To be more precise, in order for the tarball of the latest hook to be
unpacked, in the standalone build, we provide bootstrap_toolchain.py
with the location of the tarball of the latest hook based on the result
returned from the BuildDB, whereas in the CDP build we provide
cdp_install_cmd.sh with the location to which components.ini downloaded
the tarball. In either case, the unpacked tarball would be stored in the
directory of $ATLAS_HOOK_DIR, which is used by setup_build_context.py to
prepare the jar files of the hook when the docker images are built. As a
result, we always include the latest hook in the docker images.

Testing:
- Verified in the standalone build that the version of
  jackson-databind is the same as the version that fixed the issue in
  CDPD-10218, i.e., 2.9.10.4.
- Verified that this patch could be successfully built and passes the
  Quanta L0 tests in the CDP build with the fix on the cdpd side via the
  Canary tool.
- Verified that this patch passes the core tests in the DEBUG build
  except for 4 unrelated E2E tests.

Change-Id: I6d2a339645cbaa38c3a008fd646cabc4512ca226
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/79960
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
CDH-Build: Jenkins User <jenkins@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>
Quasar-L0: quasar-precommit User