CDPD-10983: Fix data load of functional.materialized_view
This is a downstream only quick fix to unblock the data load.
Recent changes in Hive broke the creation of this materialized view.
The root cause is not known yet, see the Jira for details.

Simply changing the order of the table creations seems to solve the
issue.

Change-Id: Ifdc675b05085b116dead83b489d71712a6d5195b
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/80283
CDH-Build: Csaba Ringhofer <csringhofer@cloudera.com>
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>
Quasar-L0: Csaba Ringhofer <csringhofer@cloudera.com>
Tested-by: Csaba Ringhofer <csringhofer@cloudera.com>