IMPALA-9753: Fix TRUNCATE of ACID tables on S3
The use of HDFS API was incorrect when creating an empty file
in the new base dir during truncate. Simply calling Create(Path)
does create the file in HDFS, but it is only created on S3 when
the returned stream is closed.

Testing:
- Acid truncate tests are not running on S3 as they need a running
  Hive server. Aded a regression test that will run on S3 too. It
  would be nice to run all tests on S3, but this is out of the scope
  of this change.

Change-Id: I96d315638b669c5c7198a8e47939cb2b236e35bb
Reviewed-on: http://gerrit.cloudera.org:8080/15940
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/83915
CDH-Build: Jenkins User <jenkins@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>
Quasar-L0: quasar-precommit User
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>