IMPALA-9756: avoid coord cancellation/unregistration races
This patch aims to make the query tear-down path more predictable.

Specifically, we want to ensure that Coordinator::ComputeQuerySummary()
finishes before the query is unregistered. Before that could happen
if cancellation occurred on a different thread from unregistration,
e.g. if async_exec_thread_ cancels the query, if an asynchronous
Cancel() RPC comes in, if it was cancelled internally because of
a timeout, or for some other reason. The race was possible for
a couple of reasons:
* The synchronization between threads happened on UpdateExecState(),
  which is called before ComputeQuerySummary().
* The unregistering thread may not call Coordinator::Cancel() if
  coord_exec_called_ is not set.

This patch addresses both problems, by adding synchronization
so that Coordinator::Cancel() blocks until ComputeQuerySummary()
finishes, and so that coord_exec_called_ is set even if
'async_exec_thread_' is cancelling the query (which ensures
that both threads call Coordinator::Cancel() and synchronize there).

This is only done on the finalizing thread. Other threads that
initiate cancellation should not block on cancellation, so that
threads cannot pile up waiting for the query.

Testing:
Added a regression test that failed before this change.

Ran exhaustive tests.

Change-Id: I62cacc06d9877d33b79a33aeb3b82195e639b5c4
Reviewed-on: http://gerrit.cloudera.org:8080/15954
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit 86e611bdb2a912e283ce1db01bf2940ee1583ff3)
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/85702
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
CDH-Build: Jenkins User <jenkins@cloudera.com>
Quasar-L0: Jenkins User <jenkins@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>